refactor: Login 및 Logout 관련 Controller에 @Controller 적용
