+++ app/src/main/java/com/techcourse/controller/AuthController.java
@@ -1,0 +1,63 @@
+package com.techcourse.controller;
+
+import com.techcourse.domain.User;
+import com.techcourse.repository.InMemoryUserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@Controller
+public class AuthController {
+
+    private static final ModelAndView REDIRECT_401_RESPONSE = new ModelAndView(new JspView("redirect:/401.jsp"));
+    private static final ModelAndView REDIRECT_INDEX_RESPONSE = new ModelAndView(new JspView("redirect:/index.jsp"));
+
+    private static final Logger log = LoggerFactory.getLogger(AuthController.class);
+
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView login(final HttpServletRequest req, final HttpServletResponse res) {
+        if (UserSession.isLoggedIn(req.getSession())) {
+            return REDIRECT_INDEX_RESPONSE;
+        }
+
+        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
+                .map(user -> {
+                    log.info("User : {}", user);
+                    return login(req, user);
+                })
+                .orElse(REDIRECT_401_RESPONSE);
+    }
+
+    private ModelAndView login(final HttpServletRequest request, final User user) {
+        if (user.checkPassword(request.getParameter("password"))) {
+            final var session = request.getSession();
+            session.setAttribute(UserSession.SESSION_KEY, user);
+            return REDIRECT_INDEX_RESPONSE;
+        } else {
+            return REDIRECT_401_RESPONSE;
+        }
+    }
+
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public ModelAndView loginView(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return REDIRECT_INDEX_RESPONSE;
+                })
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
+
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public ModelAndView logout(final HttpServletRequest req, final HttpServletResponse res) {
+        final var session = req.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return new ModelAndView(new JspView("redirect:/"));
+    }
+}
